+++ packages/react-vapor/src/components/index.ts
@@ -38,6 +38,7 @@ export * from './input';
 export * from './itemBox';
 export * from './labeledValue';
 export * from './lastUpdated';
+export * from './limit';
 export * from './listBox';
 export * from './loading';
 export * from './logoCard';