+++ packages/react-vapor/src/components/searchBar/index.ts
@@ -2,3 +2,4 @@ export * from './SearchBar';
 export * from './SearchBarActions';
 export * from './SearchBarConnected';
 export * from './SearchBarReducers';
+export * from './SearchBarSelector';